Update FileCache.php
Please be careful with merges and review the code better, this is not the only place that Doctrine has these errors, I just wanted to give a signal about these problems. Accept my request or not but be careful with mergem further on, I've been fixing this code for 2 days now in my own libraries. There are other files also that contain these errors.